Use Long (instead of Integer) for number of total results
The number of total search results can easily be larger than the maximum
value of Integer (2,147,483,647) in big Elasticsearch setups.

(cherry picked from commit 7e67ad94f868ca79806ff6d33b290f378d2729ff)

+++ jest-common/src/main/java/io/searchbox/core/SearchResult.java
@@ -190,10 +190,10 @@ public class SearchResult extends JestResult {
         return retval;
     }
 
-    public Integer getTotal() {
-        Integer total = null;
+    public Long getTotal() {
+        Long total = null;
         JsonNode obj = getPath(PATH_TO_TOTAL);
-        if (obj != null) total = obj.asInt();
+        if (obj != null) total = obj.asLong();
         return total;
     }
